Clase Api

+++ WebDeveloper/Areas/Personnel/Controllers/AddressController.cs
@@ -1,22 +1,14 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Web;
 using System.Web.Mvc;
-using WebDeveloper.Filters;
 using WebDeveloper.Model;
 using WebDeveloper.Repository;
-
 
 namespace WebDeveloper.Areas.Personnel.Controllers
 {
     public class AddressController : PersonBaseController<Address>
     {
-        // GET: AddressArea/Address
-        public AddressController(IRepository<Address> repository)
-            :base(repository)
+        public AddressController(IRepository<Address> repository): base(repository)
         {
-
         }
         public ActionResult Index()
         {
@@ -32,57 +24,53 @@ namespace WebDeveloper.Areas.Personnel.Controllers
         public ActionResult Create(Address address)
         {
             if (!ModelState.IsValid) return View(address);
-            address.rowguid = Guid.NewGuid();
-            address.ModifiedDate = DateTime.Now;
-            //address.BusinessEntityAddress = new BusinessEntity
-            //{
-            //    rowguid = address.rowguid,
-            //    ModifiedDate = DateTime.Now
-            //};
 
+            SetModifiedDate(address);
+            address.rowguid = Guid.NewGuid();
             _repository.Add(address);
             return RedirectToAction("Index");
         }
 
         public ActionResult Edit(int id)
         {
-            var address = _repository.GetById(x => x.AddressID == id);
-            if (address == null) return RedirectToAction("Index");
-            return View(address);
+            return View(_repository.GetById(x => x.AddressID == id));
+        }
+
+        public ActionResult Details(int id)
+        {
+            return View(_repository.GetById(x => x.AddressID == id));
         }
 
         [HttpPost]
         public ActionResult Edit(Address address)
         {
             if (!ModelState.IsValid) return View(address);
+
+            SetModifiedDate(address);
             _repository.Update(address);
             return RedirectToAction("Index");
         }
 
-
         public ActionResult Delete(int id)
         {
-            var address = _repository.GetById(x => x.AddressID == id);
-            if (address == null) return RedirectToAction("Index");
-            return View(address);
+            return View(_repository.GetById(x => x.AddressID == id));
         }
 
         [HttpPost]
-        public ActionResult Delete(Address address)
+        public ActionResult Delete(int? id )
         {
-            //if (!ModelState.IsValid) return View(person);
-            address = _repository.GetById(x => x.AddressID == address.AddressID);
+            if (!id.HasValue) return RedirectToAction("Index");
+
+            var address = _repository.GetById(x => x.AddressID == id);
+            if (address==null) return RedirectToAction("Index");
+
             _repository.Delete(address);
             return RedirectToAction("Index");
         }
 
-
-        public ActionResult Details(int id)
+        private void SetModifiedDate(Address address)
         {
-            var address = _repository.GetById(x => x.AddressID == id);
-            if (address == null) return RedirectToAction("Index");
-            return View(address);
-
+            address.ModifiedDate = DateTime.Now;
         }
     }
 }
